pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elt-lm"
version = "0.1.0"
description = "Elastic Looped Transformer (arXiv:2604.09168) + Intra-Loop Self-Distillation, causal-LM port in PyTorch"
requires-python = "==3.12.9"
readme = { text = "See plan file in ~/.claude/plans/", content-type = "text/plain" }
license = { text = "Apache-2.0" }
authors = [{ name = "local" }]

dependencies = [
    "torch>=2.7.0",
    "einops>=0.8.0",
    "transformers>=4.45.0",
    "safetensors>=0.4.3",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.2",
    "numpy>=1.26.0",
    "datasets>=2.20.0",
    "huggingface-hub>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-xdist>=3.5.0",
]
eval = [
    "datasets>=2.20.0",
    "jinja2>=3.1.0",
]
logging = [
    "wandb>=0.17.0",
    "tensorboard>=2.16.0",
]

[project.scripts]
elt-train = "elt_lm.train:cli"
elt-train-grpo = "elt_lm.train_grpo:cli"
elt-infer = "elt_lm.infer:cli"
elt-tokenize = "elt_lm.tokenize_data:cli"
elt-anytime = "elt_lm.eval.anytime_sweep:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/elt_lm"]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.pyright]
include = ["src", "tests"]
extraPaths = ["src"]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = "warning"
